✏ Rename types.js to prop-types.js in common
